requirements.txt
openai==1.58.1
python-dotenv==1.0.1
pypdf==5.1.0
pandas==2.2.3
pydantic==2.10.4
pydantic-core==2.27.2
PyMuPDF>=1.23.0